backend/requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
docutils==0.22
filetype==1.2.0
idna==3.10
packaging==25.0
Pygments==2.19.2
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
spotipy==2.25.1
urllib3==2.5.0
ytmusicapi==1.11.0
fastapi
uvicorn
